requirements.txt
tensorflow>=2.8.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
numpy>=1.21.0
pandas>=1.3.0
xgboost>=1.5.0
joblib>=1.1.0
Pillow>=9.0.0